# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "URGE v2 API",
        "description": "Prompt registry and version control API. Manage prompts, versions, render templates, store LLM results, and collaborate via teams.",
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "https://urge-next.acordado.org/api/v1"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "API key with 'urge_' prefix"
            }
        },
...
    },
    "paths": {
...
        "/prompts/{username}/{promptSlug}/branches": {
            "get": {
                "operationId": "listBranches",
                "summary": "List branches for a prompt",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "description": "User slug (owner of the prompt)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "promptSlug",
                        "in": "path",
                        "required": true,
                        "description": "Prompt slug (unique per user)",
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of branches with version counts and HEAD info"
                    }
                }
            },
            "post": {
                "operationId": "createBranch",
                "summary": "Create a new branch",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "description": "User slug (owner of the prompt)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "promptSlug",
                        "in": "path",
                        "required": true,
                        "description": "Prompt slug (unique per user)",
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "description": "Branch name (will be slugified)"
                                    },
                                    "from_version": {
                                        "type": "integer",
                                        "description": "Version number to fork from"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created"
                    }
                }
            }
        },
        "/prompts/{username}/{promptSlug}/branches/{branch}": {
            "get": {
                "operationId": "showBranch",
                "summary": "Get branch details with HEAD content",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "description": "User slug (owner of the prompt)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "promptSlug",
                        "in": "path",
                        "required": true,
                        "description": "Prompt slug (unique per user)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details"
                    }
                }
            },
            "delete": {
                "operationId": "deleteBranch",
                "summary": "Delete a non-default branch",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "description": "User slug (owner of the prompt)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "promptSlug",
                        "in": "path",
                        "required": true,
                        "description": "Prompt slug (unique per user)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted"
                    },
                    "422": {
                        "description": "Cannot delete default branch"
                    }
                }
            }
        },
        "/prompts/{username}/{promptSlug}/branches/{branch}/default": {
            "patch": {
                "operationId": "setDefaultBranch",
                "summary": "Set branch as the default",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "description": "User slug (owner of the prompt)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "promptSlug",
                        "in": "path",
                        "required": true,
                        "description": "Prompt slug (unique per user)",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default branch updated"
                    }
                }
            }
        },
...
    }
}